fifth commit

+++ src/main/java/com/example/MedicineInventoryManagement/dto/MedicineRequestDto.java
@@ -1,5 +1,6 @@
 package com.example.MedicineInventoryManagement.dto;
 
+import io.swagger.models.auth.In;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -8,17 +9,17 @@ import java.util.Date;
 @Getter
 @Setter
 public class MedicineRequestDto {
-    private Long medicineId;
+    private Integer medicineId;
     private String medicineName;
-    private Long categoryId;
+    private Integer categoryId;
     private String manufacturer;
     private String formula;
     private Date mfdDate;
     private Date expiryDate;
     private String dosage;
-    private Long quantity;
+    private Short quantity;
     private String unit;
-    private Long totalQuantity;
-    private Long costPerUnit;
+    private Integer totalQuantity;
+    private Integer costPerUnit;
     private Boolean isInStock;
 }
