+++ src/main/java/com/example/MedicineInventoryManagement/entity/Medicine.java
@@ -18,18 +18,18 @@ public class Medicine {
     @Id
     @GenericGenerator(name = "id_seq", strategy = "increment")
     @GeneratedValue(generator = "id_seq", strategy = GenerationType.AUTO)
-    private Long medicineId;
+    private Integer medicineId;
     private String medicineName;
-    private Long categoryId;
+    private Integer categoryId;
     private String manufacturer;
     private String formula;
     private Date mfdDate;
     private Date expiryDate;
     private String dosage;
-    private Long quantity;
+    private Short quantity;
     private String unit;
-    private Long totalQuantity;
-    private Long costPerUnit;
+    private Integer totalQuantity;
+    private Integer costPerUnit;
     private Boolean isInStock;
 
 //    @ManyToOne(cascade = CascadeType.ALL)
