+++ src/main/java/com/example/MedicineInventoryManagement/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package com.example.MedicineInventoryManagement.repository;
 import com.example.MedicineInventoryManagement.entity.MedicineOrder;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends CrudRepository<MedicineOrder,Long> {
+public interface OrderRepository extends CrudRepository<MedicineOrder,Integer> {
 }
