+++ src/main/java/com/example/MedicineInventoryManagement/repository/RequestRepository.java
@@ -3,5 +3,5 @@ package com.example.MedicineInventoryManagement.repository;
 import com.example.MedicineInventoryManagement.entity.Request;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RequestRepository extends CrudRepository<Request,Long> {
+public interface RequestRepository extends CrudRepository<Request,Integer> {
 }
