+++ src/main/java/com/example/MedicineInventoryManagement/service/RequestService.java
@@ -11,9 +11,8 @@ public interface RequestService {
 
     List<RequestResponseDto> getRequestList();
 
-    RequestResponseDto updateRequest(Long requestId, RequestRequestDto requestRequestDto);
+    RequestResponseDto updateRequest(Integer requestId, RequestRequestDto requestRequestDto);
 
-    RequestResponseDto deleteRequest(Long requestId);
-//    RequestResponseDto getRequestById(Long requestId);
-//    RequestResponseDto deleteRequestById(Long requestId);
+    RequestResponseDto deleteRequest(Integer requestId);
+
 }
